+++ internal/declcfg/declcfg_to_model.go
@@ -3,8 +3,8 @@ package declcfg
 import (
 	"fmt"
 
-	"github.com/operator-framework/operator-registry/pkg/model"
-	"github.com/operator-framework/operator-registry/pkg/property"
+	"github.com/operator-framework/operator-registry/internal/model"
+	"github.com/operator-framework/operator-registry/internal/property"
 )
 
 func ConvertToModel(cfg DeclarativeConfig) (model.Model, error) {
