+++ internal/declcfg/load.go
@@ -11,7 +11,7 @@ import (
 	"github.com/operator-framework/api/pkg/operators"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 
-	"github.com/operator-framework/operator-registry/pkg/property"
+	"github.com/operator-framework/operator-registry/internal/property"
 )
 
 func LoadDir(configDir string) (*DeclarativeConfig, error) {
