+++ internal/declcfg/load_test.go
