+++ internal/declcfg/testdata/invalid/not-json-object.json
