+++ internal/declcfg/testdata/invalid/not-json.txt
