+++ internal/declcfg/testdata/valid/cockroachdb.json
