+++ internal/declcfg/testdata/valid/etcd.json
