+++ internal/declcfg/testdata/valid/unrecognized-schema.json
