+++ internal/declcfg/write_test.go
